pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kalbot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "websockets>=12.0",
    "aiosqlite>=0.20",
    "xgboost>=2.0",
    "scikit-learn>=1.4",
    "pydantic>=2.0",
    "fastapi>=0.111",
    "uvicorn>=0.29",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "tomli>=2.0; python_version < '3.11'",
    "numpy>=1.26",
    "pandas>=2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.hatch.build.targets.wheel]
packages = ["src/kalbot"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
